Revert "Added generator page, routing"
