+++ package.json
@@ -11,14 +11,7 @@
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4",
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/material": "^5.6.1",
-    "bootstrap": "^5.1.3",
-    "react-bootstrap": "^2.2.3",
-    "react-router-dom": "^6.3.0",
-    "text-image": "^0.7.0"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
