--- src/components/Generator.css
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Alice&family=Barrio&family=Bungee+Shade&family=Kodchasan:wght@300&family=Lexend+Mega:wght@900&family=Major+Mono+Display&family=Megrim&family=Montserrat:wght@100&family=Notable&family=Red+Hat+Mono:wght@500&family=Rubik+Mono+One&family=Seaweed+Script&display=swap');
